refactor cli daemon targeting
