+++ src/cli/commands/insights.ts
@@ -4,7 +4,8 @@
 
 import { Command } from 'commander';
 import * as path from 'path';
-import { addDaemonHostOption, createCliDaemonClient, initCli, exitWithError } from '../utils';
+import { initCli, exitWithError } from '../utils';
+import { withCliDaemonTarget } from '../targeting';
 import { fileExists } from '../../core/services/FileService';
 import { generateInsights, formatInsightsReport } from '../../core/services/InsightsService';
 import { analyzeInsights } from '../../core/services/InsightsAnalyzer';
@@ -13,57 +14,57 @@ import { getAgent } from '../../core/codeAgents';
 import { getErrorMessage } from '../../core/utils';
 
 export function registerInsightsCommand(program: Command): void {
-    addDaemonHostOption(program
+    program
         .command('insights <session-name>')
         .description('Generate conversation insights for a session')
-        .option('--json', 'Output as JSON'))
+        .option('--json', 'Output as JSON')
         .action(async (sessionName: string, options) => {
             try {
                 const { config, repoRoot } = await initCli();
+                await withCliDaemonTarget(repoRoot, options, {
+                    daemon: async ({ client }) => {
+                        const result = await client.getSessionInsights(sessionName, {
+                            includeAnalysis: options.json,
+                        });
 
-                if (options.host) {
-                    const client = await createCliDaemonClient(repoRoot, options);
-                    const result = await client.getSessionInsights(sessionName, {
-                        includeAnalysis: options.json,
-                    });
+                        if (options.json) {
+                            console.log(JSON.stringify(result, null, 2));
+                            return;
+                        }
 
-                    if (options.json) {
-                        console.log(JSON.stringify(result, null, 2));
-                        return;
-                    }
+                        console.log(result.insights);
+                    },
+                    local: async () => {
+                        const worktreesFolder = config.get('lanes', 'worktreesFolder', '.worktrees');
+                        const worktreePath = path.join(repoRoot, worktreesFolder, sessionName);
 
-                    console.log(result.insights);
-                    return;
-                }
+                        if (!await fileExists(worktreePath)) {
+                            exitWithError(`Session '${sessionName}' not found.`);
+                        }
 
-                const worktreesFolder = config.get('lanes', 'worktreesFolder', '.worktrees');
-                const worktreePath = path.join(repoRoot, worktreesFolder, sessionName);
+                        const agentName = await getSessionAgentName(worktreePath);
+                        const agent = getAgent(agentName);
+                        if (!agent?.supportsFeature('insights')) {
+                            exitWithError(`Insights are not supported by ${agent?.displayName ?? agentName}.`);
+                        }
 
-                if (!await fileExists(worktreePath)) {
-                    exitWithError(`Session '${sessionName}' not found.`);
-                }
+                        const insights = await generateInsights(worktreePath);
 
-                const agentName = await getSessionAgentName(worktreePath);
-                const agent = getAgent(agentName);
-                if (!agent?.supportsFeature('insights')) {
-                    exitWithError(`Insights are not supported by ${agent?.displayName ?? agentName}.`);
-                }
+                        if (insights.sessionCount === 0) {
+                            console.log(`No conversation data found for session '${sessionName}'.`);
+                            return;
+                        }
 
-                const insights = await generateInsights(worktreePath);
+                        if (options.json) {
+                            console.log(JSON.stringify(insights, null, 2));
+                            return;
+                        }
 
-                if (insights.sessionCount === 0) {
-                    console.log(`No conversation data found for session '${sessionName}'.`);
-                    return;
-                }
-
-                if (options.json) {
-                    console.log(JSON.stringify(insights, null, 2));
-                    return;
-                }
-
-                const analysis = analyzeInsights(insights);
-                const report = formatInsightsReport(sessionName, insights, analysis);
-                console.log(report);
+                        const analysis = analyzeInsights(insights);
+                        const report = formatInsightsReport(sessionName, insights, analysis);
+                        console.log(report);
+                    },
+                });
             } catch (err) {
                 exitWithError(getErrorMessage(err));
             }
