+++ src/cli/commands/workflow.ts
@@ -5,7 +5,8 @@
 import { Command } from 'commander';
 import * as path from 'path';
 import * as fsPromises from 'fs/promises';
-import { addDaemonHostOption, createCliDaemonClient, initCli, exitWithError, getPackageRoot } from '../utils';
+import { initCli, exitWithError, getPackageRoot } from '../utils';
+import { withCliDaemonTarget } from '../targeting';
 import { discoverWorkflows, loadWorkflowTemplateFromString, WorkflowValidationError } from '../../core/workflow';
 import { BLANK_WORKFLOW_TEMPLATE } from '../../core/services/WorkflowService';
 import { getErrorMessage } from '../../core/utils';
@@ -15,63 +16,63 @@ export function registerWorkflowCommand(program: Command): void {
         .command('workflow')
         .description('Manage workflow templates');
 
-    addDaemonHostOption(workflow
+    workflow
         .command('list')
         .description('List available workflow templates')
-        .option('--json', 'Output as JSON'))
+        .option('--json', 'Output as JSON')
         .action(async (options) => {
             try {
                 const { config, repoRoot } = await initCli();
+                await withCliDaemonTarget(repoRoot, options, {
+                    daemon: async ({ client }) => {
+                        const response = await client.listWorkflows();
+                        const templates = response.workflows;
 
-                if (options.host) {
-                    const client = await createCliDaemonClient(repoRoot, options);
-                    const response = await client.listWorkflows();
-                    const templates = response.workflows;
+                        if (options.json) {
+                            console.log(JSON.stringify(templates, null, 2));
+                            return;
+                        }
 
-                    if (options.json) {
-                        console.log(JSON.stringify(templates, null, 2));
-                        return;
-                    }
+                        if (templates.length === 0) {
+                            console.log('No workflow templates found.');
+                            return;
+                        }
 
-                    if (templates.length === 0) {
-                        console.log('No workflow templates found.');
-                        return;
-                    }
+                        console.log(`${'NAME'.padEnd(25)} ${'SOURCE'.padEnd(12)} DESCRIPTION`);
+                        console.log('-'.repeat(70));
+                        for (const template of templates) {
+                            const source = template.isBuiltin ? 'built-in' : 'custom';
+                            console.log(`${template.name.padEnd(25)} ${source.padEnd(12)} ${template.description || ''}`);
+                        }
+                    },
+                    local: async () => {
+                        const customWorkflowsFolder = config.get('lanes', 'customWorkflowsFolder', '.lanes/workflows');
 
-                    console.log(`${'NAME'.padEnd(25)} ${'SOURCE'.padEnd(12)} DESCRIPTION`);
-                    console.log('-'.repeat(70));
-                    for (const template of templates) {
-                        const source = template.isBuiltin ? 'built-in' : 'custom';
-                        console.log(`${template.name.padEnd(25)} ${source.padEnd(12)} ${template.description || ''}`);
-                    }
-                    return;
-                }
+                        const templates = await discoverWorkflows({
+                            extensionPath: getPackageRoot(),
+                            workspaceRoot: repoRoot,
+                            customWorkflowsFolder,
+                        });
 
-                const customWorkflowsFolder = config.get('lanes', 'customWorkflowsFolder', '.lanes/workflows');
+                        if (options.json) {
+                            console.log(JSON.stringify(templates, null, 2));
+                            return;
+                        }
 
-                const templates = await discoverWorkflows({
-                    extensionPath: getPackageRoot(),
-                    workspaceRoot: repoRoot,
-                    customWorkflowsFolder,
+                        if (templates.length === 0) {
+                            console.log('No workflow templates found.');
+                            console.log(`Create one in ${customWorkflowsFolder}/`);
+                            return;
+                        }
+
+                        console.log(`${'NAME'.padEnd(25)} ${'SOURCE'.padEnd(12)} DESCRIPTION`);
+                        console.log('-'.repeat(70));
+                        for (const template of templates) {
+                            const source = template.isBuiltIn ? 'built-in' : 'custom';
+                            console.log(`${template.name.padEnd(25)} ${source.padEnd(12)} ${template.description || ''}`);
+                        }
+                    },
                 });
-
-                if (options.json) {
-                    console.log(JSON.stringify(templates, null, 2));
-                    return;
-                }
-
-                if (templates.length === 0) {
-                    console.log('No workflow templates found.');
-                    console.log(`Create one in ${customWorkflowsFolder}/`);
-                    return;
-                }
-
-                console.log(`${'NAME'.padEnd(25)} ${'SOURCE'.padEnd(12)} DESCRIPTION`);
-                console.log('-'.repeat(70));
-                for (const t of templates) {
-                    const source = t.isBuiltIn ? 'built-in' : 'custom';
-                    console.log(`${t.name.padEnd(25)} ${source.padEnd(12)} ${t.description || ''}`);
-                }
             } catch (err) {
                 exitWithError(getErrorMessage(err));
             }
